+++ src/main/resources/META-INF/plugin.xml
@@ -84,14 +84,14 @@
     <xml.elementDescriptorProvider
       implementation="com.axellience.vuegwtplugin.codeinsight.tags.VueGWTElementDescriptorProvider"/>
 
-    <fileType name="VueGWT" implementationClass="com.axellience.vuegwtplugin.language.htmltemplate.VueGWTTemplateFileType"
+    <fileType name="VueGWT" implementationClass="com.axellience.vuegwtplugin.language.VueGWTTemplateFileType"
               fieldName="INSTANCE" language="VueGWT-Template" extensions=""/>
 
     <html.embeddedContentSupport
-            implementation="com.axellience.vuegwtplugin.language.htmltemplate.VuegGWTTemplateEmbeddedContentSupport" />
+            implementation="com.axellience.vuegwtplugin.language.VuegGWTTemplateEmbeddedContentSupport" />
 
     <lang.parserDefinition language="VueGWT-Template"
-                           implementationClass="com.axellience.vuegwtplugin.language.htmltemplate.VueGWTTemplateParserDefinition"/>
+                           implementationClass="com.axellience.vuegwtplugin.language.VueGWTTemplateParserDefinition"/>
 
     <lang.syntaxHighlighterFactory language="VueGWT-Template"
                                    implementationClass="com.intellij.lang.html.HtmlSyntaxHighlighterFactory"/>
